+++ phpcs-server/src/strings.ts
@@ -31,6 +31,8 @@ export class StringResources {
 	static readonly PhpcbfUnfixableIssues: string = 'PHPCBF: Some issues could not be automatically fixed.';
 	static readonly PhpcbfError: string = '[PHPCBF] Error: {0}';
 	static readonly PhpcbfErrorMessage: string = 'PHPCBF: {0}';
+	static readonly PhpcbfDisabledWarning: string = 'PHPCBF is disabled. Enable phpcs.phpcbfEnable to use Format Document.';
+	static readonly PhpcbfExecutableNotFoundWarning: string = 'PHPCBF executable not found. Please set phpcs.phpcbfExecutablePath or ensure phpcbf is alongside phpcs.';
 
 	static readonly UnknownExecutionError: string = 'Unknown error ocurred. Please verify that {0} returns a valid json object.';
 	static readonly CodingStandardNotInstalledError: string = 'The "{0}" coding standard is not installed. Please review your configuration an try again.';